Maven 63 improve invoice functionality

+++ src/main/java/com/cydeo/accounting_app/dto/InvoiceDTO.java
@@ -5,11 +5,6 @@ import com.cydeo.accounting_app.enums.InvoiceType;
 import java.time.LocalDate;
 import java.math.BigDecimal;
 import lombok.*;
-import org.hibernate.validator.constraints.Range;
-
-import javax.validation.constraints.DecimalMin;
-import javax.validation.constraints.Digits;
-import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 @Getter
@@ -28,7 +23,6 @@ public class InvoiceDTO {
     @NotNull()
     private ClientVendorDTO clientVendor;
     private BigDecimal price;
-    private Integer tax;
+    private BigDecimal tax;
     private BigDecimal total;
-
 }
